consolidate client in dbos.go
